git-svn-id: http://192.168.20.11/sequoiadb/trunk@11877 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/cls/clsCatalogAgent.cpp
@@ -3049,7 +3049,7 @@ namespace engine
       UINT32 i = 0 ;
       while ( i++ < 4 )
       {
-         hashValue |= ( digest[i] << ( 32 - 8 * i ) ) ;
+         hashValue |= ( (UINT32)digest[i] << ( 32 - 8 * i ) ) ;
       }
       return (INT32)( hashValue >> ( 32 - partitionBit ) ) ;
    }
